trainer/requirements.txt
google-cloud-storage
google-cloud-bigquery[pandas]
pandas
numpy
scikit-learn
xgboost